Rename addControl to addPanel
