+++ examples/add.group.html
@@ -28,7 +28,7 @@ Icons generated using: http://mapicons.nicolasmollet.com/
 			addMarkers();
 
 
-			map.addControl({
+			map.addPanel({
 				template: '<div class="actions">' +
 				'<div class="places">Click to toggle Markers visibility Group: Places</div>' +
 				'<div class="parks">Click to toggle Markers visibility Group: Parks</div>' +
@@ -50,7 +50,7 @@ Icons generated using: http://mapicons.nicolasmollet.com/
 			});
 
 
-			map.addControl({
+			map.addPanel({
 				templateURL: 'code.snippet/add.group.snippet.html',
 				position: 'right center'
 			});
