+++ examples/add.marker.html
@@ -19,7 +19,7 @@
       addMarker();
 
 	    // snippet panel
-	    map.addControl({
+	    map.addPanel({
 		    templateURL: 'code.snippet/add.marker.snippet.html',
 		    position:'right center'
 	    });
