+++ examples/add.multiple.markers.html
@@ -19,7 +19,7 @@
       addMarkers();
 
 	    // snippet panel
-	    map.addControl({
+	    map.addPanel({
 		    templateURL: 'code.snippet/add.multiple.markers.snippet.html',
 		    position:'right center'
 	    });
