+++ examples/animate.markers.html
@@ -20,7 +20,7 @@
       addMarkers();
 
 	    // snippet panel
-	    map.addControl({
+	    map.addPanel({
 		    templateURL: 'code.snippet/animate.markers.snippet.html',
 		    position:'right center'
 	    });
