+++ examples/bubble.html
@@ -20,7 +20,7 @@
     if (!err) {
       addMarker();
 
-	    map.addControl({
+	    map.addPanel({
 		    template: '<div class="actions">' +
 		    '<div class="update">Update Bubble text to: "Click Me!"</div>' +
 		    '</div>',
@@ -34,7 +34,7 @@
 	    });
 
 
-	    map.addControl({
+	    map.addPanel({
 		    templateURL: 'code.snippet/add.bubble.snippet.html',
 		    position:'right center'
 	    });
