+++ examples/code.snippet/add.control.snippet.html
@@ -1,7 +1,7 @@
 <div class="map-panel" >
   <pre class="language-javascript">
 		<code class="language-javascript">
-			map.addControl({
+			map.addPanel({
         templateURL: 'templates/custom.panel.html',
         position:'left bottom',
         events: {
@@ -12,7 +12,7 @@
           }
 				});
 
-			map.addControl({
+			map.addPanel({
 		    templateURL: 'templates/snippet.panel.html',
 		    position:'right center'
 			});
