+++ examples/geojson.html
@@ -22,7 +22,7 @@
     });
 
 	  // snippet panel
-	  map.addControl({
+	  map.addPanel({
 		  templateURL: 'code.snippet/geojson.snippet.html',
 		  position:'right center'
 	  });
