+++ examples/simple.map.html
@@ -20,7 +20,7 @@
     console.log('Map Loaded!', instance);
 
 	  // snippet panel
-	  map.addControl({
+	  map.addPanel({
 		  templateURL: 'code.snippet/simple.map.snippet.html',
 		  position:'right center'
 	  });
