+++ examples/topojson.html
@@ -33,7 +33,7 @@
       }]);
     });
 
-	  map.addControl({
+	  map.addPanel({
 		  template: '<div class="actions">' +
 		  '<div class="single">Update Single Feature (Select Nevada & Colorado)</div>' +
 		  '<div class="choro">Show CensusArea Data (Chropleth) </div>' +
@@ -56,7 +56,7 @@
 
 
 	  // snippet panel
-	  map.addControl({
+	  map.addPanel({
 		  templateURL: 'code.snippet/topojson.snippet.html',
 		  position:'right center'
 	  });
