+++ lib/map-tools/addPanel.js
