+++ lib/map-tools/index.js
@@ -14,7 +14,7 @@ module.exports = function (global) {
     this.addTopoJson = require('map-tools/addFeature')(global, that).addTopoJson;
     this.addGeoJson = require('map-tools/addFeature')(global, that).addGeoJson;
     this.updateFeature = require('map-tools/updateFeature')(global, that);
-    this.addControl = require('map-tools/addControl')(global, that);
+    this.addPanel = require('map-tools/addPanel')(global, that);
     this.updateMarker = require('map-tools/updateMarker')(global, that);
     this.addGroup = require('map-tools/groups')(global, that).addGroup;
     this.updateGroup = require('map-tools/groups')(global, that).updateGroup;
